requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.2.1
Fabric==1.10.0
python-social-auth==0.2.13
